cast when printing int tokens

+++ src/libponyc/ast/token.c
@@ -127,7 +127,8 @@ const char* token_print(token_t* token)
       if (token->printed == NULL)
         token->printed = (char*)pool_alloc_size(64);
 
-      snprintf(token->printed, 64, "%llu", token->integer.low);
+      snprintf(token->printed, 64, "%llu",
+        (unsigned long long)token->integer.low);
       return token->printed;
 
     case TK_FLOAT:
